Review feedback, remove reloading warnings, move selector functions to separate TSX file
