+++ airflow/ui/src/components/FlexibleForm/FlexibleFormFieldAdvancedArray.tsx
@@ -21,12 +21,8 @@ import { githubLight, githubDark } from "@uiw/codemirror-themes-all";
 import CodeMirror from "@uiw/react-codemirror";
 
 import { useColorMode } from "src/context/colorMode";
-import type { ParamSchema } from "src/queries/useDagParams";
 
 import type { FlexibleFormElementProps } from ".";
-
-export const isFieldAdvancedArray = (fieldType: string, fieldSchema: ParamSchema) =>
-  fieldType === "array" && fieldSchema.items?.type !== "string";
 
 export const FlexibleFormFieldAdvancedArray = ({ name, param }: FlexibleFormElementProps) => {
   const { colorMode } = useColorMode();
